factor out IncomeList and IncomeTotals class from IncomeListSection
